Personal Finance Calculator

+++ python_introduction/finance_calculator.py
@@ -1,7 +1,7 @@
 ############################################################################
 
-monthly_income = int(input('Enter your monthly income: '))
-total_monthly_expenses = int(input('Enter your total monthly expenses: '))
+monthly_income = float(input('Enter your monthly income: '))
+total_monthly_expenses = float(input('Enter your total monthly expenses: '))
 
 # Calculate monthly income and monthly savings
 monthly_savings = monthly_income - total_monthly_expenses
